src/requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
ConfigArgParse==1.5.3
dataclasses==0.8
Flask==2.0.2
Flask-API==3.0.post1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.8.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
locust==2.5.0
MarkupSafe==2.0.1
msgpack==1.0.2
peewee==3.14.8
pkg_resources==0.0.0
psutil==5.8.0
psycopg2==2.9.1
PyYAML==5.4.1
pyzmq==22.3.0
redis==3.5.3
requests==2.26.0
roundrobin==0.0.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
